final norminette: ALL NORMINETTED
