+++ source/parsing/parsing_utils.c
@@ -21,6 +21,15 @@ t_redirect	*find_last(t_redirect *redir)
 	return (redir);
 }
 
+t_arg	*find_last_arg(t_arg *arg)
+{
+	if (!arg)
+		return (NULL);
+	while (arg->next)
+		arg = arg->next;
+	return (arg);
+}
+
 t_node	*create_node(t_node_type type)
 {
 	t_node	*node;
@@ -42,40 +51,4 @@ t_redir_type	get_redirect_type(t_type token_type)
 		return (R_OUT);
 	else
 		return (R_ERR);
-}
-
-int	is_builtin(char *cmd)
-{
-	int	len;
-
-	if (!cmd)
-		return (FALSE);
-	len = ft_strlen(cmd);
-	if (!ft_strncmp(cmd, "echo", len) || !ft_strncmp(cmd, "pwd", len))
-		return (TRUE);
-	else if (!ft_strncmp(cmd, "env", len) || !ft_strncmp(cmd, "cd", len))
-		return (TRUE);
-	else if (!ft_strncmp(cmd, "export", len) || !ft_strncmp(cmd, "unset", len))
-		return (TRUE);
-	else if (!ft_strncmp(cmd, "exit", len))
-		return (TRUE);
-	return (FALSE);
-}
-
-int	is_assignment(char *input)
-{
-	int	i;
-
-	i = 0;
-	if (!ft_strrchr(input, '='))
-		return (FALSE);
-	while (input[i])
-	{
-		if (input[i] == '=')
-			break ;
-		if (!is_bash_valid(input[i]))
-			return (FALSE);
-		i++;
-	}
-	return (TRUE);
 }
